+++ PSChar.h
@@ -21,7 +21,7 @@ class PSChar : public Fighter
 
 		// defensa
 		virtual void setDefensa(int);
-		virtual double getDefensa();
+		virtual int getDefensa();
 
 		// habilidad especial
 		virtual void setTieneTesoro(bool);
